+++ src/main/java/de/monticore/triggeredstatecharts/TriggeredStatechartsTool.java
@@ -1,8 +1,7 @@
 /* (c) https://github.com/MontiCore/monticore */
-package de.monticore.umlstatecharts;
+package de.monticore.triggeredstatecharts;
 
 import com.google.common.collect.Lists;
-import de.monticore.TransformationScript;
 import de.monticore.cd4code.prettyprint.CD4CodeFullPrettyPrinter;
 import de.monticore.cdbasis._ast.ASTCDClass;
 import de.monticore.class2mc.Class2MCResolver;
@@ -11,14 +10,9 @@ import de.monticore.generating.GeneratorSetup;
 import de.monticore.generating.templateengine.GlobalExtensionManagement;
 import de.monticore.generating.templateengine.TemplateController;
 import de.monticore.generating.templateengine.TemplateHookPoint;
-import de.monticore.generating.templateengine.reporting.Reporting;
-import de.monticore.generating.templateengine.reporting.commons.ASTNodeIdentHelper;
-import de.monticore.generating.templateengine.reporting.commons.ReportManager;
-import de.monticore.generating.templateengine.reporting.commons.ReportingRepository;
-import de.monticore.generating.templateengine.reporting.reporter.*;
 import de.monticore.io.paths.MCPath;
 import de.monticore.prettyprint.IndentPrinter;
-import de.monticore.prettyprint.UMLStatechartsFullPrettyPrinter;
+import de.monticore.prettyprint.TriggeredStatechartsFullPrettyPrinter;
 import de.monticore.sc2cd.HookPointService;
 import de.monticore.sc2cd.SC2CDConverter;
 import de.monticore.sc2cd.SC2CDData;
@@ -31,29 +25,16 @@ import de.monticore.scbasis._ast.ASTSCArtifact;
 import de.monticore.scbasis._ast.ASTSCState;
 import de.monticore.scbasis._cocos.*;
 import de.monticore.scbasis._symboltable.SCStateSymbol;
-import de.monticore.scevents._cocos.NonCapitalEventNames;
-import de.monticore.scevents._cocos.NonCapitalParamNames;
-import de.monticore.scevents._symboltable.SCEventsSTCompleter;
 import de.monticore.scstatehierarchy.HierarchicalStateCollector;
 import de.monticore.scstatehierarchy.NoSubstatesHandler;
 import de.monticore.symbols.basicsymbols.BasicSymbolsMill;
 import de.monticore.symbols.oosymbols.OOSymbolsMill;
-import de.monticore.tf.runtime.ODRule;
-import de.monticore.types.DeriveSymTypeOfUMLStatecharts;
-import de.monticore.types.SynthesizeSymType;
-import de.monticore.types.check.TypeCheck;
-import de.monticore.umlstatecharts._cocos.UMLStatechartsCoCoChecker;
-import de.monticore.umlstatecharts._symboltable.IUMLStatechartsArtifactScope;
-import de.monticore.umlstatecharts._symboltable.UMLStatechartsScopesGenitorDelegator;
-import de.monticore.umlstatecharts._visitor.UMLStatechartsTraverser;
+import de.monticore.triggeredstatecharts._cocos.TriggeredStatechartsCoCoChecker;
+import de.monticore.triggeredstatecharts._symboltable.ITriggeredStatechartsArtifactScope;
+import de.monticore.triggeredstatecharts._visitor.TriggeredStatechartsTraverser;
 import de.se_rwth.commons.logging.Log;
-import groovy.lang.Binding;
-import groovy.lang.GroovyShell;
 import org.apache.commons.cli.*;
-import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.codehaus.groovy.control.CompilerConfiguration;
-import org.codehaus.groovy.control.customizers.ImportCustomizer;
 
 import java.io.File;
 import java.io.IOException;
@@ -64,12 +45,12 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
-  
+public class TriggeredStatechartsTool extends TriggeredStatechartsToolTOP {
+
   public static void main(String[] args) {
-    UMLStatechartsCLI cli = new  UMLStatechartsCLI();
-    cli.init();
-    cli.run(args);
+    TriggeredStatechartsTool tool = new TriggeredStatechartsTool();
+    tool.init();
+    tool.run(args);
   }
 
   @Override
@@ -98,9 +79,9 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
       // we need the global scope for symbols and cocos
       MCPath symbolPath = new MCPath(Paths.get(""));
       if (cmd.hasOption("path")) {
-        symbolPath = new MCPath(Arrays.stream(cmd.getOptionValues("path")).map(x -> Paths.get(x)).collect(Collectors.toList()));
+        symbolPath = new MCPath(Arrays.stream(cmd.getOptionValues("path")).map(Paths::get).collect(Collectors.toList()));
       }
-      UMLStatechartsMill.globalScope().setSymbolPath(symbolPath);
+      TriggeredStatechartsMill.globalScope().setSymbolPath(symbolPath);
       BasicSymbolsMill.initializePrimitives();
       Class2MCResolver resolver = new Class2MCResolver();
       OOSymbolsMill.globalScope().addAdaptedOOTypeSymbolResolver(resolver);
@@ -108,17 +89,12 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
 
       // parse input file, which is now available
       // (only returns if successful)
-      String input = cmd.getOptionValue("i");
-      ASTSCArtifact scartifact = parse(input);
+      ASTSCArtifact scartifact = parse(cmd.getOptionValue("i"));
 
-      IUMLStatechartsArtifactScope scope = createSymbolTable(scartifact);
+      ITriggeredStatechartsArtifactScope scope = createSymbolTable(scartifact);
 
       // check context conditions
       runDefaultCoCos(scartifact);
-
-      if (cmd.hasOption("t")) {
-        doTrafos(scartifact, cmd.getOptionValues("t"));
-      }
 
       if (cmd.hasOption("s")) {
         String path = cmd.getOptionValue("s", StringUtils.EMPTY);
@@ -139,87 +115,18 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
 
       // -option generate to CD
       if (cmd.hasOption("gen")) {
-        String path = cmd.getOptionValue("gen");
+        String path = cmd.getOptionValue("gen", StringUtils.EMPTY);
         String configTemplate = cmd.getOptionValue("ct", StringUtils.EMPTY);
         String templatePath = cmd.getOptionValue("fp", StringUtils.EMPTY);
         String handcodedPath = cmd.getOptionValue("hcp", StringUtils.EMPTY);
-        String reportDirectory = cmd.getOptionValue("genr", path + "/report");
 
-        // Setup reporting
-        ReportManager.ReportManagerFactory factory = getReportManagerFactory(path, reportDirectory);
-        Reporting.init(path, reportDirectory, factory);
-
-        String modelName = FilenameUtils.getBaseName(input);
-        Reporting.on(modelName);
-        Reporting.reportModelStart(scartifact, modelName, "");
-        Reporting.reportParseInputFile(new File(input).getAbsoluteFile().toPath(), modelName);
-
-        // Generate the CD
         generateCD(scartifact, path, configTemplate, templatePath, handcodedPath);
-
-        // Flush reporting
-        Reporting.reportModelEnd(modelName, "");
-        Reporting.flush(scartifact);
       }
-
 
     } catch (ParseException e) {
       // ann unexpected error from the apache CLI parser:
-      Log.error("0xA5C02 Could not process CLI parameters: " + e.getMessage());
+      Log.error("0xA5C04 Could not process CLI parameters: " + e.getMessage());
     }
-  }
-
-  protected ReportManager.ReportManagerFactory getReportManagerFactory(String outputDirectory, String reportDirectory) {
-    return new ReportManager.ReportManagerFactory() {
-      @Override public ReportManager provide(String modelName) {
-        String lowerCaseName = modelName.toLowerCase();
-        ReportManager reports = new ReportManager(outputDirectory);
-        ReportingRepository repository = new ReportingRepository(new ASTNodeIdentHelper());
-        repository.initAllTemplates();
-
-        SummaryReporter summary = new SummaryReporter(reportDirectory, lowerCaseName, repository, UMLStatechartsMill.inheritanceTraverser());
-        GeneratedFilesReporter generated = new GeneratedFilesReporter(reportDirectory, lowerCaseName, repository);
-        HandWrittenCodeReporter handwritten = new HandWrittenCodeReporter(reportDirectory, lowerCaseName, repository);
-        TemplatesReporter templates = new TemplatesReporter(reportDirectory, lowerCaseName, repository);
-        TransformationReporter transformations = new TransformationReporter(reportDirectory, lowerCaseName, repository);
-        SuccessfulReporter finishReporter = new SuccessfulReporter(reportDirectory, lowerCaseName);
-        IncGenCheckReporter incGenCheck = new IncGenCheckReporter(reportDirectory, lowerCaseName);
-        IncGenGradleReporter gradleReporter = new IncGenGradleReporter(reportDirectory, lowerCaseName);
-
-        reports.addReportEventHandler(summary); // 01_Summary
-        reports.addReportEventHandler(generated); // 02_GeneratedFiles
-        reports.addReportEventHandler(handwritten); // 03_HandwrittenCodeFiles
-        reports.addReportEventHandler(templates); // 04_Templates
-        reports.addReportEventHandler(transformations); // 14_Transformations
-        reports.addReportEventHandler(finishReporter); // 19_Successful
-        reports.addReportEventHandler(incGenCheck); // IncGenCheck
-        reports.addReportEventHandler(gradleReporter);
-
-        return reports;
-      }
-    };
-  }
-
-  /**
-   * Creates the symbol table from the parsed AST.
-   *
-   * @param ast The top statechart model element.
-   * @return The artifact scope derived from the parsed AST
-   */
-  @Override
-  public IUMLStatechartsArtifactScope createSymbolTable(ASTSCArtifact ast) {
-
-    // create scope and symbol skeleton
-    UMLStatechartsScopesGenitorDelegator genitor = UMLStatechartsMill.scopesGenitorDelegator();
-    IUMLStatechartsArtifactScope symTab = genitor.createFromAST(ast);
-
-    // complete symbols including type check
-    UMLStatechartsTraverser completer = UMLStatechartsMill.traverser();
-    TypeCheck typeCheck = new TypeCheck(new SynthesizeSymType(),new DeriveSymTypeOfUMLStatecharts());
-    completer.add4SCEvents(new SCEventsSTCompleter(typeCheck));
-    ast.accept(completer);
-
-    return symTab;
   }
 
   /**
@@ -248,7 +155,7 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
   public static final String REPORT_STATE_NAMES = "stateNames.txt";
 
   public String reportReachableStates(ASTSCArtifact ast) {
-    UMLStatechartsTraverser traverser = UMLStatechartsMill.traverser();
+    TriggeredStatechartsTraverser traverser = TriggeredStatechartsMill.traverser();
     // collect all states
     // HierarchicalStateCollector vs StateCollector
     HierarchicalStateCollector stateCollector = new HierarchicalStateCollector();
@@ -256,10 +163,10 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
     traverser.add4SCStateHierarchy(stateCollector);
     ast.accept(traverser);
     Set<String> statesToBeChecked = stateCollector.getStates()
-      .stream().map(e -> e.getName()).collect(Collectors.toSet());
+      .stream().map(ASTSCState::getName).collect(Collectors.toSet());
 
     // collect all initial states
-    traverser = UMLStatechartsMill.traverser();
+    traverser = TriggeredStatechartsMill.traverser();
     InitialStateCollector initialStateCollector = new InitialStateCollector();
     traverser.add4SCBasis(initialStateCollector);
     //  only find real initial states
@@ -275,7 +182,7 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
       String from = currentlyChecked.iterator().next();
       currentlyChecked.remove(from);
       ReachableStateCollector reachableStateCollector = new ReachableStateCollector(from);
-      traverser = UMLStatechartsMill.traverser();
+      traverser = TriggeredStatechartsMill.traverser();
       traverser.add4SCBasis(reachableStateCollector);
       ast.accept(traverser);
       for (String to : reachableStateCollector.getReachableStates()) {
@@ -310,7 +217,7 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
 
   public String reportBranchingDegree(ASTSCArtifact ast) {
     BranchingDegreeCalculator branchingDegreeCalculator = new BranchingDegreeCalculator();
-    UMLStatechartsTraverser traverser = UMLStatechartsMill.traverser();
+    TriggeredStatechartsTraverser traverser = TriggeredStatechartsMill.traverser();
     traverser.add4SCBasis(branchingDegreeCalculator);
     ast.accept(traverser);
     return branchingDegreeCalculator.getBranchingDegrees().entrySet().stream()
@@ -320,11 +227,11 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
 
   public String reportStateNames(ASTSCArtifact ast) {
     StateCollector stateCollectorVisitor = new StateCollector();
-    UMLStatechartsTraverser traverser = UMLStatechartsMill.traverser();
+    TriggeredStatechartsTraverser traverser = TriggeredStatechartsMill.traverser();
     traverser.add4SCBasis(stateCollectorVisitor);
     ast.accept(traverser);
     return String.join(", ", stateCollectorVisitor.getStates()
-      .stream().map(e -> e.getName()).collect( Collectors.toSet())) + System.lineSeparator();
+      .stream().map(ASTSCState::getName).collect( Collectors.toSet())) + System.lineSeparator();
   }
 
   /**
@@ -334,44 +241,17 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
    */
   @Override
   public void runDefaultCoCos(ASTSCArtifact ast) {
-    UMLStatechartsCoCoChecker checker = new UMLStatechartsCoCoChecker();
+    TriggeredStatechartsCoCoChecker checker = new TriggeredStatechartsCoCoChecker();
     checker.addCoCo(new UniqueStates());
     checker.addCoCo(new TransitionSourceTargetExists());
-    UMLStatechartsTraverser t = UMLStatechartsMill.traverser();
+    TriggeredStatechartsTraverser t = TriggeredStatechartsMill.traverser();
     t.setSCStateHierarchyHandler(new NoSubstatesHandler());
     checker.addCoCo(new AtLeastOneInitialState(t));
     checker.addCoCo(new CapitalStateNames());
     checker.addCoCo(new PackageCorrespondsToFolders());
     checker.addCoCo(new SCFileExtension());
     checker.addCoCo(new SCNameIsArtifactName());
-    checker.addCoCo(new NonCapitalEventNames());
-    checker.addCoCo(new NonCapitalParamNames());
     checker.checkAll(ast);
-  }
-
-  /**
-   * Apply transformation groovy workflow scripts
-   * @param ast The SC-AST to be transformed
-   * @param trafoScripts the array of trafos groovy script to be applied
-   */
-  public void doTrafos(ASTSCArtifact ast, String[] trafoScripts) {
-    CompilerConfiguration config = new CompilerConfiguration();
-    config.setScriptBaseClass(TransformationScript.class.getName()); // Groovy base script providing trafo helpers
-    // By default, import all trafos from the default de.monticore.tf package
-    config.addCompilationCustomizers(new ImportCustomizer().addStarImports("de.monticore.tf"));
-
-    GroovyShell shell = new GroovyShell(this.getClass().getClassLoader(), new Binding(), config);
-
-    for (String script : trafoScripts) {
-      try {
-        shell.getContext().setVariable("ast", ast);
-        TransformationScript groovyScript = (TransformationScript) shell.parse(new File(script));
-        groovyScript.__setTransformationGroovyShell(shell); // Pass this shell to respect imports
-        groovyScript.run();
-      }catch (IOException e){
-        throw new RuntimeException(e.getMessage(), e.getCause());
-      }
-    }
   }
 
   /**
@@ -384,8 +264,8 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
   @Override
   public void prettyPrint(ASTSCArtifact scartifact, String file) {
     // pretty print AST
-    UMLStatechartsFullPrettyPrinter prettyPrinterDelegator
-      = new UMLStatechartsFullPrettyPrinter();
+    TriggeredStatechartsFullPrettyPrinter prettyPrinterDelegator
+      = new TriggeredStatechartsFullPrettyPrinter();
     String prettyOutput = prettyPrinterDelegator.prettyprint(scartifact);
     print(prettyOutput, file);
   }
@@ -402,11 +282,7 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
    * @param outputDirectory The target directory name for outputting the CD artifact. If empty,
    *          the content is printed to stdout instead
    */
-  public void generateCD(ASTSCArtifact scartifact,
-                         String outputDirectory,
-                         String configTemplate,
-                         String templatePath,
-                         String handcodedPath) {
+  public void generateCD(ASTSCArtifact scartifact, String outputDirectory, String configTemplate, String templatePath, String handcodedPath) {
     // pretty print AST
     SC2CDConverter converter = new SC2CDConverter();
 
@@ -424,7 +300,7 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
       setup.setTracing(false);
     }
 
-    SC2CDData sc2CDData = converter.doConvertUML(scartifact, setup);
+    SC2CDData sc2CDData = converter.doConvertTriggered(scartifact, setup);
 
     if (!handcodedPath.isEmpty()) {
       SCTopDecorator topDecorator = new SCTopDecorator(new MCPath(handcodedPath));
@@ -456,14 +332,14 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
         templateSetup.setGlex(glex);
 
         if (!templatePath.isEmpty()) {
-          List<File> files = Lists.newArrayList();
+          List<File>  files = Lists.newArrayList();
           try (Stream<Path> paths = Files.walk(Paths.get(templatePath))) {
             paths
-                .filter(Files::isRegularFile)
-                .forEach(f -> files.add(f.toFile()));
+              .filter(Files::isRegularFile)
+              .forEach(f -> files.add(f.toFile()));
           }
           catch (IOException e) {
-            Log.error("0x5C701 Incorrect template path "+ templatePath);
+            Log.error("0x5C700 Incorrect template path "+ templatePath);
           }
         }
 
@@ -471,27 +347,24 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
         TemplateHookPoint hp = new TemplateHookPoint(configTemplate);
         // Provide the glex, a template helper, and the CD classes as args
         List<Object> args = Arrays.asList(templateSetup.getGlex(),
-                                          new HookPointService(),
-                                          sc2CDData.getScClass(),
-                                          sc2CDData.getStateSuperClass(),
-                                          sc2CDData.getStateClasses());
-        String qualifiedTemplateName = configTemplate.replaceAll(".ftl","");
-        Reporting.reportTemplateStart(qualifiedTemplateName, sc2CDData.getCompilationUnit());
+          new HookPointService(),
+          sc2CDData.getScClass(),
+          sc2CDData.getStateSuperClass(),
+          sc2CDData.getStateClasses());
         hp.processValue(tc, sc2CDData.getCompilationUnit(), args);
-        Reporting.reportTemplateEnd(qualifiedTemplateName, sc2CDData.getCompilationUnit());
       }
 
       for (ASTCDClass clazz : sc2CDData.getCompilationUnit().getCDDefinition().getCDClassesList()) {
         Path out = Paths.get(packageDir.toString(), clazz.getName() + ".java");
         generatorEngine.generate("de.monticore.sc2cd.gen.Class", out,
-                                 clazz, printer, sc2CDData.getCompilationUnit().getCDPackageList());
+          clazz, printer, sc2CDData.getCompilationUnit().getCDPackageList());
       }
 
     }
   }
 
   /**
-   * Initializes the standard CLI options for the Statechart tool.
+   * Initializes the standard options for the Statechart tool.
    *
    * @return The CLI options with arguments.
    */
@@ -547,7 +420,7 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
   }
 
   /**
-   * Initializes the additional CLI options for the Statechart tool.
+   * Initializes the additional options for the Statechart tool.
    *
    * @return The CLI options with arguments.
    */
@@ -555,63 +428,47 @@ public class UMLStatechartsCLI extends UMLStatechartsCLITOP {
   public Options addAdditionalOptions(Options options) {
     // convert to state pattern CD
     options.addOption(Option.builder("gen")
-        .longOpt("generate")
-        .argName("dir")
-        .optionalArg(true)
-        .numberOfArgs(1)
-        .desc("Prints the state pattern CD-AST to stdout or the generated java classes to the specified folder (optional)")
-        .build());
+      .longOpt("generate")
+      .argName("dir")
+      .optionalArg(true)
+      .numberOfArgs(1)
+      .desc("Prints the state pattern CD-AST to stdout or the generated java classes to the specified folder (optional)")
+      .build());
 
     // configTemplate parameter
     options.addOption(Option.builder("ct")
-        .longOpt("configTemplate")
-        .argName("file")
-        .optionalArg(true)
-        .numberOfArgs(1)
-        .desc("Provides a config template (optional)")
-        .build());
+      .longOpt("configTemplate")
+      .argName("file")
+      .optionalArg(true)
+      .numberOfArgs(1)
+      .desc("Provides a config template (optional)")
+      .build());
 
     // handcoded gen path parameter
     options.addOption(Option.builder("hcp")
-        .longOpt("handcodedPath")
-        .argName("pathlist")
-        .optionalArg(true)
-        .numberOfArgs(1)
-        .desc("List of directories to look for handwritten code to integrate (optional)")
-        .build());
+      .longOpt("handcodedPath")
+      .argName("pathlist")
+      .optionalArg(true)
+      .numberOfArgs(1)
+      .desc("List of directories to look for handwritten code to integrate (optional)")
+      .build());
 
 
     // templatePath parameter
     options.addOption(Option.builder("fp")
-        .longOpt("templatePath")
-        .argName("pathlist")
-        .optionalArg(true)
-        .numberOfArgs(1)
-        .desc("List of directories to look for handwritten templates to integrate (optional)")
-        .build());
+      .longOpt("templatePath")
+      .argName("pathlist")
+      .optionalArg(true)
+      .numberOfArgs(1)
+      .desc("List of directories to look for handwritten templates to integrate (optional)")
+      .build());
 
     // model paths
     options.addOption(Option.builder("path")
-        .argName("pathlist")
-        .hasArgs()
-        .desc("Sets the artifact path for imported symbols, space separated.")
-        .build());
-
-    // trafo script parameter
-    options.addOption(Option.builder("t")
-                              .longOpt("trafo")
-                              .argName("groovyscript")
-                              .hasArgs()
-                              .desc("Specifies the path for a groovy script applying transformations (optional)")
-                              .build());
-
-    // specify generate reports path
-    options.addOption(Option.builder("genr")
-                              .longOpt("genreport")
-                              .argName("path")
-                              .hasArgs()
-                              .desc("Specifies the directory for printing reports based on the given model.")
-                              .build());
+      .argName("pathlist")
+      .hasArgs()
+      .desc("Sets the artifact path for imported symbols, space separated.")
+      .build());
 
     return options;
   }